backend/requirements.txt
Flask==2.0.2
Flask-CORS==3.0.10
Flask-PyMongo==2.3.0
python-dotenv==0.19.2
tensorflow==2.7.0
Werkzeug==2.0.2

